// src/app/shared/models/co2-table.model.ts
export interface Co2TableModel {
  origin?: string;
  destination?: string;
  item1?: string;
  item2?: string;
  item3?: string;
  item4?: string;
  item5?: string;
  item6?: string;
  item7?: string;
  item8?: string;
  item9?: string;
  item10?: string;
  item11?: string;
  item12?: string;
  item13?: string;
  item14?: string;
  item15?: string;
  item16?: string;
}

export interface Co2TableHeaderModel {
  origin: string;
  destination: string;
  item1: string;
  item2: string;
  item3: string;
  item4: string;
  item5: string;
  item6: string;
  item7: string;
  item8: string;
  item9: string;
  item10: string;
  item11: string;
  item12: string;
  item13: string;
  item14: string;
  item15: string;
  item16: string;
}